Fixes for issues revealed through ClangCL compilation
Fix std::move being unqualified in a whole bunch of places; remove
 using std::move where it is no longer needed because std::move is
 supposed to be qualified.
Specify explicitly that auto-destructors are not to be created and used
 with stream buffer and string stream pointers, because these
 "exit-destructors" can cause segfaults because the type destructor
 handles dismantling and destroying these unique_ptr objects.
Refactoring problematic names of constants.
Fixed missing EOF newlines.
Fixed implicitly casting INT_MIN to unsigned.

+++ src/TestCPPUtil.cpp
@@ -92,7 +92,7 @@ namespace TestCPP {
             }
 
             if (toCast >= static_cast<unsigned>(INT_MIN)) {
-                return static_cast<int>(toCast - INT_MIN) + INT_MIN;
+                return static_cast<int>(toCast - static_cast<unsigned>(INT_MIN)) + INT_MIN;
             }
 
             throw std::runtime_error(
